refactor(security): extract security subsystem into ts-security crate (#359)
Вынос подсистемы безопасности из монолита `src-tauri/src/security/` в крейт `ts-security`:
API-валидаторы (api_validator/api_validator_service), OAuth (oauth_handler), импорт .env
(env_importer), команды (commands/additional_commands), registry. Хранилище ключей уже было
в `ts-secure-storage` (#91) — здесь остаётся шим-модуль.

Разблокировано выносом ts-core (#361):
- crate::core::{AppEvent, EventBus, Service} -> ts_core::{AppEvent, EventBus, Service}
- crate::video_compiler::error -> ts_render::video_compiler::error
- crate::video_compiler::VideoCompilerState -> ts_render_services::VideoCompilerState
- crate::command_registry -> ts_command_registry; crate::security:: (intra) -> crate::

Монолит: `security/` -> файл-шим `src/security.rs = pub use ts_security::*` (state, analysis,
app_builder резолвятся без правок).
Зависимости: ts-core, ts-render, ts-render-services, ts-secure-storage, ts-command-registry,
tauri, reqwest, url, anyhow, chrono, uuid, async-trait.

Проверки: cargo test -p ts-security --test-threads=2 -> 116 passed;
cargo check --lib (src-tauri) -> 0 ошибок, 3 baseline-варнинга.

Closes #359. Часть #354.
