+++ crates/ts-security/src/secure_storage.rs
@@ -1,4 +1,4 @@
 //! Зашифрованное хранилище API-ключей — вынесено в крейт `ts-secure-storage` (#91).
-//! Ре-экспорт-шим: `crate::security::secure_storage::*` и `security/mod.rs::pub use secure_storage::*`
+//! Ре-экспорт-шим: `crate::secure_storage::*` и `security/mod.rs::pub use secure_storage::*`
 //! резолвятся без правок (vision_commands / state / ai_api_proxy и др.).
 pub use ts_secure_storage::*;
