+++ crates/ts-security/src/tests/secure_storage_test.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod tests {
-  use crate::security::secure_storage::*;
+  use crate::secure_storage::*;
   use std::str::FromStr;
 
   #[test]
