+++ src-tauri/Cargo.toml
@@ -92,6 +92,7 @@ ts-platform = { path = "../crates/ts-platform" }
 ts-analysis = { path = "../crates/ts-analysis" }
 ts-events = { path = "../crates/ts-events" }
 ts-secure-storage = { path = "../crates/ts-secure-storage" }
+ts-security = { path = "../crates/ts-security" }
 ts-render = { path = "../crates/ts-render" }
 ts-render-services = { path = "../crates/ts-render-services" }
 ts-render-analysis = { path = "../crates/ts-render-analysis" }
